+++ optionspilot/data/__init__.py
@@ -1,5 +1,7 @@
 from optionspilot.data.base import MarketDataProvider
 from optionspilot.data.cache import CandleCache
+from optionspilot.data.cached import CachedProvider
 from optionspilot.data.yfinance_provider import YFinanceProvider
 
-__all__ = ["MarketDataProvider", "CandleCache", "YFinanceProvider"]
+__all__ = ["MarketDataProvider", "CandleCache", "CachedProvider",
+           "YFinanceProvider"]